WIP: Add support for Bittrex exchange via API
- Fix time.parsing with variable millisecond length
- Add withdrawal support
